% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btable.R
\name{btable}
\alias{btable}
\title{Create a balance table}
\usage{
btable(
  data,
  id.vars,
  bal.vars,
  group.var,
  panel = F,
  new.names.bal.vars = NULL,
  save.path = NULL
)
}
\arguments{
\item{data}{Data set to be used. Can be a panel (one row per observation-period), with
variables to be balanced as columns if that is the case, set \code{panel} to \code{TRUE}.
Alternatively, \code{data} may be in long format, with each row a unique combination of
observation, period and variable. A column \code{value} should be specified, which will hold
the unique value of the row.}

\item{id.vars}{The identifying variables, uniquely identifying an observation and year.}

\item{bal.vars}{The variables to balance across groups.}

\item{group.var}{The variable identifying the two groups, treatment and control. Should be
binary, with 0 identifying the control group and 1 the treatment.}

\item{panel}{By default \code{F}. Switch to \code{T} is data set is a panel with variables to be
balanced as columns.}

\item{new.names.bal.vars}{A named vector used to rename the variables to be balanced.}

\item{save.path}{By default \code{NULL}. Indicate a path where the balance table should be saved.}
}
\value{
A \code{gt} table object.
}
\description{
This function creates a standard balance tables as used in economics.
It takes a data set (coerced to a data.table), identifying variables,
a group (or treatment) variable, and variables to balance across groups.
Optionally, it may save the output to a LaTeX file and change the names of
the balanced variables.
}
