% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stder}
\alias{stder}
\title{Compute standard error of a sample mean}
\usage{
stder(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector of values.}
}
\value{
The standard error as a numeric value.
}
\description{
Given a distribution, returns its standard error.
}
